services/order_service/requirements.txt
aiohttp==3.13.2
redis==7.1.0
python-dotenv==1.2.1
